Move class CancelableImpl to package org.mastodon.mamut.feature
